pyproject.toml
[tool.poetry]
name = "781-project"
version = "0.1.0"
description = "This is my 781 Project 2024"
authors = ["nathanMils <106823768+nathanMils@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
seaborn = "^0.13.2"
matplotlib = "^3.9.2"
ucimlrepo = "^0.0.7"
scipy = "^1.14.1"
scikit-learn = "^1.5.2"
ipykernel = "^6.29.5"
tabulate = "^0.9.0"
mlflow = "^2.17.2"
imblearn = "^0.0"
joblib = "^1.4.2"
python-whois = "^0.9.4"
flask = "^3.0.3"
memory-profiler = "^0.61.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
